affiliate-deals-automation/requirements.txt
requests
beautifulsoup4
python-dotenv
sqlite-utils
python-telegram-bot
schedule
